scheduling-interface/requirements.txt
grpcio>=1.68.1
grpcio-tools>=1.68.1
protobuf>=6.0.0
fastapi==0.115.6
uvicorn==0.34.0
requests==2.32.3